% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absAZIMUTH.R
\name{absAZIMUTH}
\alias{absAZIMUTH}
\title{Calculate northness/southness of a site relative to another (from latitude longitude)}
\usage{
absAZIMUTH(dist, lonlats)
}
\arguments{
\item{dist}{This is a distance matrix between two sites based on lonlats}

\item{lonlats}{a dataframe containing columns entited Site, Lon, Lat, Pop}
}
\value{
A distance dist in kilometres between all pairs sites
}
\description{
Calculate the absolute azimuth between two sites
}
\examples{
dta <- data.frame(Site= LETTERS[1:4], Lat= 1:4, Lon= 5:8, Pop=100:103)
dist <- point2DIST(dta)
absAZIMUTH(diat=dist, lonlats=dta)

}
