% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSUPERNODE.R
\name{addSUPERNODE}
\alias{addSUPERNODE}
\title{Add supersource and supersink nodes}
\usage{
addSUPERNODE(dist, sources, sinks)
}
\arguments{
\item{dist}{This is a distance matrix between two sites}

\item{sources}{vector of site names to be added as sources from the supersource node}

\item{sinks}{vector of site names to be added as sinks to the supersink node}
}
\value{
A distance matrix in kilometres between all pairs sites
}
\description{
These nodes will later act as source and sink nodes which initiate and terminate the population flow through the network. More specifically, they are added to the distance matrix with values of zero (for the time being)
}
\examples{
dta <- data.frame(Site= LETTERS[1:4], Lat= 1:4, Lon= 5:8, Pop=100:103)
dist <- point2DIST(dta)
addSUPERNODE(dist)



}
