% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directedNET.R
\name{directedNET}
\alias{directedNET}
\title{Only allow the network to flow from source to sink}
\usage{
directedNET(dist, include_diagonal = TRUE, replaceWith = 0)
}
\arguments{
\item{dist}{This is a distance matrix between two sites}

\item{include_diagonal}{`TRUE` means the diagonal is replaced (such that animals cannot remain in a site) and `FALSE` means they are not}

\item{replaceWith}{Replace lower part of distance matrix with e.g. zeros or NAs, default is set to 0}
}
\value{
A distance matrix in kilometres between all pairs sites
}
\description{
This function replaces the bottom side of the distance network with zeroes. This ensures that everything can only go in one direction. i.e. from code example A to B, A to C, A to D, B to C, B to D and C to D
}
\examples{
dta <- data.frame(Site= LETTERS[1:4], Lat= 1:4, Lon= 5:8, Pop=100:103)
dist <- point2DIST(dta)
directedNET(dist)


}
