% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxflow.R
\docType{package}
\name{maxflow}
\alias{maxflow}
\alias{maxflow-package}
\title{maxflow}
\description{
This package calculates the maximum flow through a network using integel linear programming.
}
\author{
Kiran Dhanjal-Adams \email{kiran.dhanjal.adams@gmail.com}
}
