% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodePopPROP.R
\name{nodePopPROP}
\alias{nodePopPROP}
\title{Calculate the proportion of population that can flow into each site}
\usage{
nodePopPROP(dta, population = 1000)
}
\arguments{
\item{dta}{This is a dataframe. Columns of interest must be named: Site, Lat (for latitude), Lon (for longitude).}

\item{population}{This is the total size of the populaiton or number of individuals allowed to flow through the network}
}
\value{
A matrix containing the proportion of the population able to flow into that
}
\description{
This function calculates determines the proportion of the population flowing into a site - this is used to later parameterise preference
}
\examples{
dta <- data.frame(Site= LETTERS[1:4], Lat= 1:4, Lon= 5:8, Pop=100:103)
nodePopPROP(dta, 300)



}
