% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point2DIST.R
\name{point2DIST}
\alias{point2DIST}
\title{Make a distance matrix}
\usage{
point2DIST(dta, miles = FALSE,
  proj4string = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs ")
}
\arguments{
\item{dta}{This is a dataframe. Columns of interest must be named: Site, Lat (for latitude), Lon (for longitude).}
}
\value{
A distance matrix in kilometres between all pairs sites
}
\description{
Make a distance matrix
}
\examples{
dta <- data.frame(Site= LETTERS[1:4], Lat= 1:4, Lon= 5:8, Pop=100:103)
point2DIST(dta)

}
