% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point_2_Distance_Network.R
\name{Point_2_Distance_Network}
\alias{Point_2_Distance_Network}
\title{Make a distance matrix}
\usage{
Point_2_Distance_Network(dta)
}
\arguments{
\item{dta}{This is a dataframe. Columns of interest must be named: Site, Lat (for latitude), Lon (for longitude).}
}
\value{
A distance matrix in kilometres between all pairs sites
}
\description{
Make a distance matrix
}
\examples{
dta <- data.frame(Site= LETTERS[1:4], Lat= 1:4, Lon= 5:8, Pop=100:103)
Point_2_Distance_Network(dta)


}
