% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popPROP.R
\name{popPROP}
\alias{popPROP}
\title{Calculate the proportion of population that can flow into each site}
\usage{
popPROP(network, population)
}
\arguments{
\item{population}{This is the total size of the populaiton or number of individuals allowed to flow through the network}

\item{dta}{This is a dataframe. Columns of interest must be named: Site, Lat (for latitude), Lon (for longitude).}
}
\value{
A matrix containing the proportion of the population able to flow into that
}
\description{
This function calculates determines the proportion of the population flowing into a site - this is used to later parameterise preference
}
\examples{
dta <- data.frame(Site= LETTERS[1:4], Lat= 1:4, Lon= 5:8, Pop=100:103)
network <- point2DIST(dta)
popPROP(network, 300)



}
