% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritiseFLOW.R
\name{prioritiseFLOW}
\alias{prioritiseFLOW}
\title{Prioritise sites according to population flow}
\usage{
prioritiseFLOW(network, sites, method = "igraph", plot = TRUE)
}
\arguments{
\item{network}{adjacency network describing where birds are going}

\item{sites}{network sites including latitude and longitude, and population count at each site}

\item{method}{max_flow method, currently that of igraph, but will also support gurobi or lpsolve}

\item{plot}{TRUE/FALSE to determine whether the output is plotted or not}
}
\value{
a list containting the prioritisation (as a list of the sites in the order they are removed), the network which was randomly generated,
the tracks that were randomly generated, and the sites that were randomly generated for animals to use.
}
\description{
Prioritise sites according to population flow
}
\examples{
pop=100000
rand_net = randomNET(nsites=15,pop=pop)

# priotise sites according to flow through network
prioritiseFLOW(rand_net$network, rand_net$sites)

}
